+++ packages/nimble-components/.storybook/main.js
@@ -39,5 +39,6 @@ module.exports = {
         };
 
         return config;
-    }
+    },
+    staticDirs: ['public']
 };